Audit input components: per-component test split + spec/a11y sync
Final category-wide audit closing the agnostic input epic.

Tests:
- Split monolithic spec_conformance/input.rs and
  proptest_state_machines/input.rs into per-component modules,
  matching the overlay/utility layout.
- Add state-machine proptests for checkbox_group and radio_group.

Drift fixes:
- radio_group error message uses aria-live="polite" (was role="alert"),
  matching its spec and the sibling input components.
- checkbox/switch/text_field/textarea now emit aria-errormessage on
  invalid, aligning all error-message-bearing input components with the
  foundation a11y convention (03-accessibility.md).
- ars-i18n fallback formatter appends the percent sign for Style::Percent,
  fixing semantically-wrong aria-valuetext ("50" -> "50%") in no-ICU4X
  builds (root cause of the meter/progress snapshot failures).

Spec sync:
- Document aria-errormessage across all 10 error-message-bearing input
  specs (code examples + ARIA tables).
- Document pin-input `dir`, search-input `debounce_pending`, and
  checkbox-group max_checked clamping; align switch spec Effect enum.

Closes #220

Entire-Checkpoint: 1e27022b2cfa
