+++ crates/ars-components/src/input/text_field/mod.rs
@@ -900,6 +900,10 @@ impl Api<'_> {
 
         if self.ctx.invalid {
             attrs.set(HtmlAttr::Aria(AriaAttr::Invalid), "true");
+            attrs.set(
+                HtmlAttr::Aria(AriaAttr::ErrorMessage),
+                self.ctx.ids.part("error-message"),
+            );
         }
 
         if let Some(placeholder) = &self.ctx.placeholder {
@@ -1395,6 +1399,10 @@ mod tests {
             Some("name-field-description name-field-error-message")
         );
         assert_eq!(attrs.get(&HtmlAttr::Aria(AriaAttr::Invalid)), Some("true"));
+        assert_eq!(
+            attrs.get(&HtmlAttr::Aria(AriaAttr::ErrorMessage)),
+            Some("name-field-error-message")
+        );
     }
 
     #[test]
