+++ crates/ars-components/tests/spec_conformance/input/mod.rs
@@ -1,0 +1,16 @@
+//! Spec-conformance tests for `crates/ars-components/src/input/*`.
+
+mod checkbox;
+mod checkbox_group;
+mod editable;
+mod file_trigger;
+mod number_input;
+mod password_input;
+mod pin_input;
+mod radio_group;
+mod range_slider;
+mod search_input;
+mod slider;
+mod switch;
+mod text_field;
+mod textarea;
