+++ crates/ars-components/tests/spec_conformance/input/textarea.rs
@@ -1,0 +1,18 @@
+use ars_components::input::textarea;
+
+use crate::helper::assert_anatomy;
+
+#[test]
+fn textarea_anatomy_matches_spec() {
+    assert_anatomy(
+        "textarea",
+        &[
+            (textarea::Part::Root, "root"),
+            (textarea::Part::Label, "label"),
+            (textarea::Part::Textarea, "textarea"),
+            (textarea::Part::CharacterCount, "character-count"),
+            (textarea::Part::Description, "description"),
+            (textarea::Part::ErrorMessage, "error-message"),
+        ],
+    );
+}
