+++ spec/components/input/number-input.md
@@ -518,7 +518,10 @@ impl<'a> Api<'a> {
             attrs.set(HtmlAttr::Aria(AriaAttr::DescribedBy), self.ctx.ids.part("error-message"));
         }
         if self.ctx.required { attrs.set(HtmlAttr::Aria(AriaAttr::Required), "true"); }
-        if self.ctx.invalid { attrs.set(HtmlAttr::Aria(AriaAttr::Invalid), "true"); }
+        if self.ctx.invalid {
+            attrs.set(HtmlAttr::Aria(AriaAttr::Invalid), "true");
+            attrs.set(HtmlAttr::Aria(AriaAttr::ErrorMessage), self.ctx.ids.part("error-message"));
+        }
         if self.ctx.disabled { attrs.set_bool(HtmlAttr::Disabled, true); }
         if self.ctx.readonly { attrs.set_bool(HtmlAttr::ReadOnly, true); }
         if let Some(ref name) = self.ctx.name { attrs.set(HtmlAttr::Name, name); }
@@ -658,17 +661,18 @@ NumberInput
 
 ### 3.1 ARIA Roles, States, and Properties
 
-| Property           | Element | Value                                    |
-| ------------------ | ------- | ---------------------------------------- |
-| `role`             | Input   | `spinbutton`                             |
-| `aria-valuenow`    | Input   | Current numeric value                    |
-| `aria-valuemin`    | Input   | Minimum value (when finite)              |
-| `aria-valuemax`    | Input   | Maximum value (when finite)              |
-| `aria-valuetext`   | Input   | Locale-formatted value string            |
-| `aria-required`    | Input   | Present when `required=true`             |
-| `aria-invalid`     | Input   | Present when `invalid=true`              |
-| `aria-labelledby`  | Input   | Points to Label id                       |
-| `aria-describedby` | Input   | Points to Description + ErrorMessage ids |
+| Property            | Element | Value                                         |
+| ------------------- | ------- | --------------------------------------------- |
+| `role`              | Input   | `spinbutton`                                  |
+| `aria-valuenow`     | Input   | Current numeric value                         |
+| `aria-valuemin`     | Input   | Minimum value (when finite)                   |
+| `aria-valuemax`     | Input   | Maximum value (when finite)                   |
+| `aria-valuetext`    | Input   | Locale-formatted value string                 |
+| `aria-required`     | Input   | Present when `required=true`                  |
+| `aria-invalid`      | Input   | Present when `invalid=true`                   |
+| `aria-errormessage` | Input   | Points to ErrorMessage id when `invalid=true` |
+| `aria-labelledby`   | Input   | Points to Label id                            |
+| `aria-describedby`  | Input   | Points to Description + ErrorMessage ids      |
 
 ### 3.2 Keyboard Interaction
