+++ spec/components/input/textarea.md
@@ -6,8 +6,8 @@ foundation_deps: [architecture, accessibility, interactions, forms]
 shared_deps: []
 related: [text-field]
 references:
-  ark-ui: Field
-  react-aria: TextField
+    ark-ui: Field
+    react-aria: TextField
 ---
 
 # Textarea
@@ -467,7 +467,10 @@ impl<'a> Api<'a> {
         if self.ctx.disabled { attrs.set_bool(HtmlAttr::Disabled, true); }
         if self.ctx.readonly { attrs.set_bool(HtmlAttr::ReadOnly, true); }
         if self.ctx.required { attrs.set_bool(HtmlAttr::Required, true); }
-        if self.ctx.invalid { attrs.set(HtmlAttr::Aria(AriaAttr::Invalid), "true"); }
+        if self.ctx.invalid {
+            attrs.set(HtmlAttr::Aria(AriaAttr::Invalid), "true");
+            attrs.set(HtmlAttr::Aria(AriaAttr::ErrorMessage), self.ctx.ids.part("error-message"));
+        }
         if let Some(p) = &self.ctx.placeholder { attrs.set(HtmlAttr::Placeholder, p); }
         if let Some(max) = self.ctx.max_length { attrs.set(HtmlAttr::MaxLength, max.to_string()); }
         if let Some(min) = self.ctx.min_length { attrs.set(HtmlAttr::MinLength, min.to_string()); }
@@ -594,12 +597,13 @@ Textarea
 
 ### 3.1 ARIA Roles, States, and Properties
 
-| Property           | Element  | Value                                    |
-| ------------------ | -------- | ---------------------------------------- |
-| `aria-invalid`     | Textarea | Present when `invalid=true`              |
-| `aria-required`    | Textarea | Implicit via `required` attribute        |
-| `aria-labelledby`  | Textarea | Points to Label id                       |
-| `aria-describedby` | Textarea | Points to Description + ErrorMessage ids |
+| Property            | Element  | Value                                         |
+| ------------------- | -------- | --------------------------------------------- |
+| `aria-invalid`      | Textarea | Present when `invalid=true`                   |
+| `aria-errormessage` | Textarea | Points to ErrorMessage id when `invalid=true` |
+| `aria-required`     | Textarea | Implicit via `required` attribute             |
+| `aria-labelledby`   | Textarea | Points to Label id                            |
+| `aria-describedby`  | Textarea | Points to Description + ErrorMessage ids      |
 
 - Character count element uses `aria-live="polite"` so screen readers announce remaining characters on change.
 - Auto-resize does not affect accessibility — no ARIA changes needed.
